% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{read_metadata}
\alias{read_metadata}
\title{Read all JSON metadata in folder}
\usage{
read_metadata(path)
}
\arguments{
\item{path}{path that contains JSON metadata files}
}
\value{
A list where each entry contains the metadata of a output.
The metadata is stored as a list.
}
\description{
Read all JSON metadata in folder
}
