% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{save_output_and_metadata}
\alias{save_output_and_metadata}
\title{Save output and metadata}
\usage{
save_output_and_metadata(
  tfl_output,
  tfl_path,
  name,
  numbering,
  title,
  subtitle,
  footnotes = NULL,
  type = "table",
  program_name,
  nonbreaking_space = (type == "table")
)
}
\arguments{
\item{tfl_output}{output can either be a data.frame, data.table, or a list of data.frames
, where each entry will be shown on it's own page. For figures provide a ggplot object.}

\item{tfl_path}{the path to the folder where files will be saved}

\item{name}{Name of the output. This will also be the file name.}

\item{title}{The title of the output. This is used for the header and the table of contents.}

\item{subtitle}{Optional subtitle}

\item{footnotes}{a character vector, where each entry will be shown on a separate line.}

\item{type}{'table', 'listing', or 'figure'}

\item{program_name}{the name of the program that generated the output.}

\item{nonbreaking_space}{a bool indicating if normal spaces should be replaced with nonbreaking spaces when calculating the column width of tables.}

\item{number}{Appendix section number of output. For example 14.2.1 for a efficacy table.}
}
\description{
This functions saves the output and the metadata
as a JSON-file. Tables and listings are saved as RDS-files, whereas
figures are saved as PNG-files. A footnote is added with the time, date and program name
for traceability.
}
\examples{
save_output_and_metadata(
  tfl_output=mtcars,
  tfl_path='myoutput',
  name='cars',
  number='1.1',
  title='Cars',
  subtitle='',
  footnotes='Source: 1974 Motor Trend US magazine',
  type='listing',
  program_name='example.r'
)

}
