% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TFL_ordering.R
\name{sort_metadata}
\alias{sort_metadata}
\title{Sort metadata by numbering}
\usage{
sort_metadata(metadata)
}
\arguments{
\item{metadata}{a list generated from read_metadata}
}
\value{
metadata ordered by `numbering` values.
}
\description{
The `numbering` value contains sections (for example 14.2.1) stored as
characters. This function sorts the metadata to make sure the numbering values
are in the correct order. This allows us to generate RTF-code for each output and then
combine them in the correct order.
}
