% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{split_into_list}
\alias{split_into_list}
\title{Split dataset into list of datasets}
\usage{
split_into_list(dataset, by_variable, values_per_page, blankout = TRUE)
}
\description{
This is so you can manually decide how to split a table into pages.
For example you could have 4 unique values of `Visit` per page.
}
\examples{
iris$Species = as.character(iris$Species)
split_into_list(iris, 'Species', 2)
}
