% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{write_rtf}
\alias{write_rtf}
\title{Creates RTF-file with output}
\usage{
write_rtf(
  metadata,
  metadata_path,
  output_filename,
  document_title,
  title_page_info,
  header_rtf = ""
)
}
\arguments{
\item{metadata}{metadata object that you get from `read_metadata`.}

\item{metadata_path}{the path to the RDS files and PNG files of the output}

\item{output_filename}{path of the resulting RTF-file}

\item{document_title}{the title of the document on the front page}

\item{title_page_info}{A data.frame with a `field` and `value` columns. This information is shown on title page.}

\item{header_rtf}{Raw RTF-code that can be added to header}
}
\description{
Creates RTF-file with output
}
